[pa-112] Refacto: move all infra ods matters into utils
