+++ api/lib/infrastructure/utils/ods/read-ods-utils.js
@@ -1,40 +1,15 @@
-const util = require('util');
-const fs = require('fs');
-const JSZip = require('jszip');
-const XLSX = require('xlsx');
+const { ODSBufferReadFailedError, ODSTableDataEmptyError, ODSTableHeadersNotFoundError } = require('../../../domain/errors');
+const { loadOdsZip } = require('./common-ods-utils');
 const _ = require('lodash');
-const { ODSBufferReadFailedError, ODSTableDataEmptyError, ODSTableHeadersNotFoundError } = require('../../domain/errors');
+const XLSX = require('xlsx');
 
 const CONTENT_XML_IN_ODS = 'content.xml';
 
-module.exports = {
-  getContentXml,
-  makeUpdatedOdsByContentXml,
-  extractTableDataFromOdsFile,
-};
-
 async function getContentXml({ odsFilePath }) {
-  const zip = await _loadOdsTemplate(odsFilePath);
+  const zip = await loadOdsZip(odsFilePath);
   const contentXmlBufferCompressed = zip.file(CONTENT_XML_IN_ODS);
   const uncompressedBuffer = await contentXmlBufferCompressed.async('nodebuffer');
   return Buffer.from(uncompressedBuffer, 'utf8').toString();
-}
-
-async function makeUpdatedOdsByContentXml({ stringifiedXml, odsFilePath }) {
-  const zip = await _loadOdsTemplate(odsFilePath);
-  await zip.file(CONTENT_XML_IN_ODS, stringifiedXml);
-  const odsBuffer = await zip.generateAsync({ type: 'nodebuffer' });
-  return odsBuffer;
-}
-
-async function _loadOdsTemplate(odsFilePath) {
-  const odsFileData = await _openOdsFile(odsFilePath);
-  const zip = JSZip();
-  return zip.loadAsync(odsFileData);
-}
-
-function _openOdsFile(odsFilePath) {
-  return util.promisify(fs.readFile)(odsFilePath);
 }
 
 async function extractTableDataFromOdsFile({ odsBuffer, tableHeaderTargetPropertyMap }) {
@@ -92,12 +67,12 @@ function _allHeadersValuesAreInTheRow(row, headers) {
 
 function _mapSheetHeadersWithProperties(sheetHeaderRow, tableHeaderTargetPropertyMap) {
   return _(sheetHeaderRow)
-    .map(addTargetDatas(tableHeaderTargetPropertyMap))
+    .map(_addTargetDatas(tableHeaderTargetPropertyMap))
     .compact()
     .value();
 }
 
-function addTargetDatas(tableHeaderTargetPropertyMap) {
+function _addTargetDatas(tableHeaderTargetPropertyMap) {
   return (header, columnName) => {
     const targetProperties = _.find(tableHeaderTargetPropertyMap, { header });
     if (targetProperties) {
@@ -118,3 +93,8 @@ function _transformSheetDataRow(sheetDataRow, sheetHeaderPropertyMap) {
     return target;
   }, {});
 }
+
+module.exports = {
+  getContentXml,
+  extractTableDataFromOdsFile,
+};
